refactored the classes
