--- pi5/scripts/config.py
@@ -1,6 +1,0 @@
-CAMERA_CONFIG = {
-    "camera_id": 0,
-    "image_width": 1280,
-    "image_height": 720,
-    "fps": 20
-}
