+++ pi5/scripts/utils/config.py
@@ -1,0 +1,12 @@
+CAMERA_CONFIG = {
+    "camera_id": 0,
+    "image_width": 1280,
+    "image_height": 720,
+    "fps": 20
+}
+
+TAG_CONFIG = {
+    "tag_size": 54,              # in millimeter
+    "small_tag_size": 10.8,      # in millimeter
+    "tag_family": "tagCustom48h12"
+}
